grocify_pipeline/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.0
Pillow>=10.1.0
jsonschema>=4.20.0

